chore : 유저의 소분류 카테고리 별 정답률 조회 api 위치 profile로 이동
