+++ cs25-service/src/main/java/com/example/cs25service/domain/profile/service/ProfileService.java
@@ -1,11 +1,14 @@
 package com.example.cs25service.domain.profile.service;
 
+import com.example.cs25entity.domain.quiz.entity.QuizCategory;
+import com.example.cs25entity.domain.quiz.repository.QuizCategoryRepository;
 import com.example.cs25entity.domain.subscription.entity.SubscriptionHistory;
 import com.example.cs25entity.domain.subscription.repository.SubscriptionHistoryRepository;
 import com.example.cs25entity.domain.user.entity.User;
 import com.example.cs25entity.domain.user.exception.UserException;
 import com.example.cs25entity.domain.user.exception.UserExceptionCode;
 import com.example.cs25entity.domain.user.repository.UserRepository;
+import com.example.cs25entity.domain.userQuizAnswer.entity.UserQuizAnswer;
 import com.example.cs25entity.domain.userQuizAnswer.repository.UserQuizAnswerRepository;
 import com.example.cs25service.domain.profile.dto.ProfileResponseDto;
 import com.example.cs25service.domain.profile.dto.ProfileWrongQuizResponseDto;
@@ -15,10 +18,13 @@ import com.example.cs25service.domain.security.dto.AuthUser;
 import com.example.cs25service.domain.subscription.dto.SubscriptionHistoryDto;
 import com.example.cs25service.domain.subscription.dto.SubscriptionInfoDto;
 import com.example.cs25service.domain.subscription.service.SubscriptionService;
+import com.example.cs25service.domain.userQuizAnswer.dto.CategoryUserAnswerRateResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
@@ -29,6 +35,7 @@ public class ProfileService {
     private final UserRepository userRepository;
     private final SubscriptionService subscriptionService;
     private final SubscriptionHistoryRepository subscriptionHistoryRepository;
+    private final QuizCategoryRepository quizCategoryRepository;
 
     // 구독 정보 가져오기
     public UserSubscriptionResponseDto getUserSubscription(AuthUser authUser) {
@@ -90,5 +97,43 @@ public class ProfileService {
                 user.getScore(),
                 myRank
         );
+    }
+
+    public CategoryUserAnswerRateResponse getUserQuizAnswerCorrectRate(Long userId){
+        //유저 검증
+        User user = userRepository.findByIdOrElseThrow(userId);
+        if(!user.isActive()){
+            throw new UserException(UserExceptionCode.INACTIVE_USER);
+        }
+
+        //유저 Id에 따른 구독 정보의 대분류 카테고리 조회
+        QuizCategory parentCategory = quizCategoryRepository.findQuizCategoryByUserId(userId);
+
+        //소분류 조회
+        List<QuizCategory> childCategories = parentCategory.getChildren();
+
+        Map<String, Double> rates = new HashMap<>();
+        //유저가 푼 문제들 중, 소분류에 속하는 로그 다 가져와
+        for(QuizCategory child : childCategories){
+            List<UserQuizAnswer> answers = userQuizAnswerRepository.findByUserIdAndQuizCategoryId(userId, child.getId());
+
+            if (answers.isEmpty()) {
+                rates.put(child.getCategoryType(), 0.0);
+                continue;
+            }
+
+            long totalAnswers = answers.size();
+            long correctAnswers = answers.stream()
+                    .filter(UserQuizAnswer::getIsCorrect) // 정답인 경우 필터링
+                    .count();
+
+            double answerRate = (double) correctAnswers / totalAnswers * 100;
+            rates.put(child.getCategoryType(), answerRate);
+
+        }
+
+        return CategoryUserAnswerRateResponse.builder()
+                .correctRates(rates)
+                .build();
     }
 }
